--- a/pkg/services/sqlstore/searchstore/builder.go
+++ b/pkg/services/sqlstore/searchstore/builder.go
@@ -35,7 +35,7 @@ func (b *Builder) ToSQL(limit, page int64) (string, []interface{}) {
 	b.sql.WriteString("\n")
 
 	b.sql.WriteString(
-		`LEFT OUTER JOIN dashboard AS folder ON folder.id = dashboard.folder_id
+		`LEFT OUTER JOIN dashboard AS folder ON folder.org_id = dashboard.org_id AND folder.id = dashboard.folder_id
 		LEFT OUTER JOIN dashboard_tag ON dashboard.id = dashboard_tag.dashboard_id`)
 	b.sql.WriteString("\n")
 	b.sql.WriteString(orderQuery)