<!--

Thank you for sending a pull request! Here are some tips:

1. If this is your first time, please read our contribution guide at https://github.com/grafana/grafana/blob/main/CONTRIBUTING.md

2. Ensure you include and run the appropriate tests as part of your Pull Request.

3. In a new feature or configuration option, an update to the documentation is necessary. Everything related to the documentation is under the docs folder in the root of the repository.

4. If the Pull Request is a work in progress, make use of GitHub's "Draft PR" feature and mark it as such.

5. If you can not merge your Pull Request due to a merge conflict, Rebase it. This gets it in sync with the main branch.

6. Name your PR as "<FeatureArea>: Describe your change", e.g. Alerting: Prevent race condition. If it's a fix or feature relevant for the changelog describe the user impact in the title. The PR title is used to auto-generate the changelog for issues marked with the "add to changelog" label.

7. If your PR content should be added to the What's New document for the next major or minor release, add the **add to what's new** label to your PR. Note that you should add this label to the main PR that introduces the feature; do not add this label to smaller PRs for the feature.

-->

**What is this feature?**

[Add a brief description of what the feature or update does.]

**Why do we need this feature?**

[Add a description of the problem the feature is trying to solve.]

**Who is this feature for?**

[Add information on what kind of user the feature is for.]

**Which issue(s) does this PR fix?**:

<!--

- Automatically closes linked issue when the Pull Request is merged.

Usage: "Fixes #<issue number>", or "Fixes (paste link of issue)"

-->

Fixes #

**Special notes for your reviewer:**

<details><summary>MySQL 5.7</summary>
<p>

```
% benchstat mysql57_logs_grafana_grafana_125182_1_14.log mysql57_logs_grafana_grafana_125193_1_14.log
goos: linux
goarch: amd64
pkg: github.com/grafana/grafana/pkg/api
cpu: Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz
                                                                                         │ mysql57_logs_grafana_grafana_125182_1_14.log │ mysql57_logs_grafana_grafana_125193_1_14.log │
                                                                                         │                    sec/op                    │       sec/op         vs base                 │
FolderListAndSearch/get_root_folders_with_nested_folders_feature_enabled-32                                              100.69m ± 3%             99.71m ± 3%        ~ (p=0.195 n=8)
FolderListAndSearch/get_subfolders_with_nested_folders_feature_enabled-32                                                 10.03m ± 9%             10.11m ± 9%        ~ (p=0.382 n=8)
FolderListAndSearch/list_all_inherited_dashboards_with_nested_folders_feature_enabled-32                                  97.228 ± 0%              4.950 ± 1%  -94.91% (p=0.000 n=8)
FolderListAndSearch/search_for_pattern_with_nested_folders_feature_enabled-32                                             96.516 ± 0%              1.882 ± 3%  -98.05% (p=0.000 n=8)
FolderListAndSearch/search_for_specific_dashboard_nested_folders_feature_enabled-32                                       96.659 ± 0%              1.806 ± 5%  -98.13% (p=0.000 n=8)
FolderListAndSearch/get_root_folders_with_nested_folders_feature_disabled-32                                              8.222m ± 1%             9.008m ± 2%   +9.55% (p=0.000 n=8)
FolderListAndSearch/list_all_dashboards_with_nested_folders_feature_disabled-32                                           109.76 ± 1%              20.01 ± 2%  -81.77% (p=0.000 n=8)
FolderListAndSearch/search_specific_dashboard_with_nested_folders_feature_disabled-32                                    109.305 ±  ∞ ¹            1.498 ± 4%        ~ (p=0.222 n=1+8)
geomean                                                                                                                    4.164                  510.0m       -87.75%
¹ need >= 6 samples for confidence interval at level 0.95

                                                                                         │ mysql57_logs_grafana_grafana_125182_1_14.log │ mysql57_logs_grafana_grafana_125193_1_14.log │
                                                                                         │                     B/op                     │        B/op          vs base                 │
FolderListAndSearch/get_root_folders_with_nested_folders_feature_enabled-32                                             6.084Mi ± 11%           5.951Mi ± 10%        ~ (p=0.442 n=8)
FolderListAndSearch/get_subfolders_with_nested_folders_feature_enabled-32                                               560.5Ki ± 40%           560.6Ki ± 48%        ~ (p=0.878 n=8)
FolderListAndSearch/list_all_inherited_dashboards_with_nested_folders_feature_enabled-32                                26.86Mi ± 15%           26.66Mi ± 15%   -0.73% (p=0.021 n=8)
FolderListAndSearch/search_for_pattern_with_nested_folders_feature_enabled-32                                           835.4Ki ±  9%           558.2Ki ±  6%  -33.18% (p=0.000 n=8)
FolderListAndSearch/search_for_specific_dashboard_nested_folders_feature_enabled-32                                     571.4Ki ±  3%           284.8Ki ±  3%  -50.17% (p=0.000 n=8)
FolderListAndSearch/get_root_folders_with_nested_folders_feature_disabled-32                                            1.092Mi ±  0%           1.092Mi ±  0%        ~ (p=0.574 n=8)
FolderListAndSearch/list_all_dashboards_with_nested_folders_feature_disabled-32                                         26.71Mi ± 15%           26.34Mi ± 15%        ~ (p=0.161 n=8)
FolderListAndSearch/search_specific_dashboard_with_nested_folders_feature_disabled-32                                  567.38Ki ±   ∞ ¹         71.47Ki ±  5%        ~ (p=0.222 n=1+8)
geomean                                                                                                                 2.250Mi                 1.506Mi        -33.09%
¹ need >= 6 samples for confidence interval at level 0.95

                                                                                         │ mysql57_logs_grafana_grafana_125182_1_14.log │ mysql57_logs_grafana_grafana_125193_1_14.log │
                                                                                         │                  allocs/op                   │      allocs/op       vs base                 │
FolderListAndSearch/get_root_folders_with_nested_folders_feature_enabled-32                                               152.4k ± 0%             152.4k ± 0%        ~ (p=0.367 n=8)
FolderListAndSearch/get_subfolders_with_nested_folders_feature_enabled-32                                                 15.65k ± 0%             15.65k ± 0%        ~ (p=0.174 n=8)
FolderListAndSearch/list_all_inherited_dashboards_with_nested_folders_feature_enabled-32                                  595.1k ± 0%             590.9k ± 0%   -0.72% (p=0.000 n=8)
FolderListAndSearch/search_for_pattern_with_nested_folders_feature_enabled-32                                            12.195k ± 3%             7.088k ± 0%  -41.88% (p=0.000 n=8)
FolderListAndSearch/search_for_specific_dashboard_nested_folders_feature_enabled-32                                       6439.0 ± 5%              564.0 ± 5%  -91.24% (p=0.000 n=8)
FolderListAndSearch/get_root_folders_with_nested_folders_feature_disabled-32                                              24.58k ± 0%             24.58k ± 0%   +0.01% (p=0.000 n=8)
FolderListAndSearch/list_all_dashboards_with_nested_folders_feature_disabled-32                                           600.8k ± 0%             592.4k ± 0%   -1.41% (p=0.000 n=8)
FolderListAndSearch/search_specific_dashboard_with_nested_folders_feature_disabled-32                                    11032.0 ±  ∞ ¹            486.0 ± 9%        ~ (p=0.222 n=1+8)
geomean                                                                                                                   45.43k                  21.14k       -53.47%
¹ need >= 6 samples for confidence interval at level 0.95
```

</p>
</details> 

<details><summary>MySQL 8</summary>
<p>

```
% benchstat mysql8_logs_grafana_grafana_125182_1_15.log mysql8_logs_grafana_grafana_125193_1_15.log
goos: linux
goarch: amd64
pkg: github.com/grafana/grafana/pkg/api
cpu: Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz
                                                                                         │ mysql8_logs_grafana_grafana_125182_1_15.log │ mysql8_logs_grafana_grafana_125193_1_15.log │
                                                                                         │                   sec/op                    │        sec/op         vs base               │
FolderListAndSearch/get_root_folders_with_nested_folders_feature_enabled-32                                                107.4m ± 7%           104.0m ±  5%        ~ (p=0.328 n=8)
FolderListAndSearch/get_subfolders_with_nested_folders_feature_enabled-32                                                  10.19m ± 5%           10.30m ±  5%        ~ (p=0.645 n=8)
FolderListAndSearch/list_all_inherited_dashboards_with_nested_folders_feature_enabled-32                                  4640.1m ± 4%           453.9m ± 17%  -90.22% (p=0.000 n=8)
FolderListAndSearch/search_for_pattern_with_nested_folders_feature_enabled-32                                               6.914 ± 8%            4.498 ±  6%  -34.94% (p=0.000 n=8)
FolderListAndSearch/search_for_specific_dashboard_nested_folders_feature_enabled-32                                         6.941 ± 8%            4.131 ±  4%  -40.49% (p=0.000 n=8)
FolderListAndSearch/get_root_folders_with_nested_folders_feature_disabled-32                                               7.926m ± 3%           8.480m ±  1%   +6.99% (p=0.000 n=8)
FolderListAndSearch/list_all_dashboards_with_nested_folders_feature_disabled-32                                           2386.4m ± 0%           344.6m ±  3%  -85.56% (p=0.000 n=8)
FolderListAndSearch/search_specific_dashboard_with_nested_folders_feature_disabled-32                                       1.987 ± 3%            3.866 ±  3%  +94.52% (p=0.000 n=8)
geomean                                                                                                                    556.2m                317.0m        -43.00%

                                                                                         │ mysql8_logs_grafana_grafana_125182_1_15.log │ mysql8_logs_grafana_grafana_125193_1_15.log │
                                                                                         │                    B/op                     │         B/op          vs base               │
FolderListAndSearch/get_root_folders_with_nested_folders_feature_enabled-32                                              5.828Mi ±  8%          6.066Mi ± 11%        ~ (p=0.328 n=8)
FolderListAndSearch/get_subfolders_with_nested_folders_feature_enabled-32                                                560.5Ki ± 39%          560.5Ki ±  0%        ~ (p=0.878 n=8)
FolderListAndSearch/list_all_inherited_dashboards_with_nested_folders_feature_enabled-32                                 24.45Mi ±  8%          25.77Mi ±  5%   +5.40% (p=0.015 n=8)
FolderListAndSearch/search_for_pattern_with_nested_folders_feature_enabled-32                                            320.3Ki ± 19%          345.6Ki ± 10%        ~ (p=0.234 n=8)
FolderListAndSearch/search_for_specific_dashboard_nested_folders_feature_enabled-32                                      86.13Ki ±  4%          78.66Ki ±  3%   -8.67% (p=0.000 n=8)
FolderListAndSearch/get_root_folders_with_nested_folders_feature_disabled-32                                             1.092Mi ±  0%          1.093Mi ±  0%        ~ (p=0.442 n=8)
FolderListAndSearch/list_all_dashboards_with_nested_folders_feature_disabled-32                                          26.26Mi ± 15%          23.58Mi ± 11%  -10.19% (p=0.010 n=8)
FolderListAndSearch/search_specific_dashboard_with_nested_folders_feature_disabled-32                                    75.09Ki ±  3%          88.55Ki ±  4%  +17.93% (p=0.000 n=8)
geomean                                                                                                                  1.201Mi                1.221Mi         +1.71%

                                                                                         │ mysql8_logs_grafana_grafana_125182_1_15.log │ mysql8_logs_grafana_grafana_125193_1_15.log │
                                                                                         │                  allocs/op                  │      allocs/op        vs base               │
FolderListAndSearch/get_root_folders_with_nested_folders_feature_enabled-32                                                152.4k ± 0%            152.4k ± 0%        ~ (p=0.204 n=8)
FolderListAndSearch/get_subfolders_with_nested_folders_feature_enabled-32                                                  15.65k ± 0%            15.65k ± 0%        ~ (p=1.000 n=8)
FolderListAndSearch/list_all_inherited_dashboards_with_nested_folders_feature_enabled-32                                   590.8k ± 0%            590.5k ± 0%   -0.04% (p=0.000 n=8)
FolderListAndSearch/search_for_pattern_with_nested_folders_feature_enabled-32                                              7.281k ± 0%            7.171k ± 0%   -1.50% (p=0.000 n=8)
FolderListAndSearch/search_for_specific_dashboard_nested_folders_feature_enabled-32                                         808.5 ± 9%             634.5 ± 6%  -21.52% (p=0.000 n=8)
FolderListAndSearch/get_root_folders_with_nested_folders_feature_disabled-32                                               24.58k ± 0%            24.58k ± 0%   +0.01% (p=0.000 n=8)
FolderListAndSearch/list_all_dashboards_with_nested_folders_feature_disabled-32                                            590.8k ± 0%            590.6k ± 0%   -0.03% (p=0.000 n=8)
FolderListAndSearch/search_specific_dashboard_with_nested_folders_feature_disabled-32                                       546.0 ± 6%             834.5 ± 5%  +52.84% (p=0.000 n=8)
geomean                                                                                                                    22.50k                 22.97k        +2.10%                                                                                                                            34.96k                        22.90k       -34.49%
```

</p>
</details> 

<details><summary>PostgreSQL</summary>
<p>

```
% benchstat baseline_postgres_logs_grafana_grafana_124456_1_13.log 124663_postgres_logs_grafana_grafana_124663_1_13.log
goos: linux
goarch: amd64
pkg: github.com/grafana/grafana/pkg/api
cpu: Intel(R) Xeon(R) Platinum 8375C CPU @ 2.90GHz
                                                                                         │ postgres_logs_grafana_grafana_125182_1_13.log │ postgres_logs_grafana_grafana_125193_1_13.log │
                                                                                         │                    sec/op                     │        sec/op          vs base                │
FolderListAndSearch/get_root_folders_with_nested_folders_feature_enabled-32                                                 102.4m ±  2%            104.2m ±  5%         ~ (p=0.328 n=8)
FolderListAndSearch/get_subfolders_with_nested_folders_feature_enabled-32                                                   10.72m ±  5%            11.07m ±  6%         ~ (p=0.721 n=8)
FolderListAndSearch/list_all_inherited_dashboards_with_nested_folders_feature_enabled-32                                    727.3m ± 14%           1668.3m ±  3%  +129.38% (p=0.000 n=8)
FolderListAndSearch/search_for_pattern_with_nested_folders_feature_enabled-32                                               640.9m ±  0%            206.1m ± 12%   -67.84% (p=0.000 n=8)
FolderListAndSearch/search_for_specific_dashboard_nested_folders_feature_enabled-32                                         651.7m ±  1%            203.5m ±  2%   -68.78% (p=0.000 n=8)
FolderListAndSearch/get_root_folders_with_nested_folders_feature_disabled-32                                                7.003m ±  3%            7.068m ±  3%         ~ (p=0.798 n=8)
FolderListAndSearch/list_all_dashboards_with_nested_folders_feature_disabled-32                                             491.3m ±  4%           1192.7m ±  1%  +142.76% (p=0.000 n=8)
FolderListAndSearch/search_specific_dashboard_with_nested_folders_feature_disabled-32                                       475.9m ±  0%            154.0m ±  1%   -67.64% (p=0.000 n=8)
geomean                                                                                                                     164.9m                  134.1m         -18.65%

                                                                                         │ postgres_logs_grafana_grafana_125182_1_13.log │ postgres_logs_grafana_grafana_125193_1_13.log │
                                                                                         │                     B/op                      │          B/op            vs base              │
FolderListAndSearch/get_root_folders_with_nested_folders_feature_enabled-32                                                6.483Mi ±  9%             6.430Mi ±  8%       ~ (p=0.878 n=8)
FolderListAndSearch/get_subfolders_with_nested_folders_feature_enabled-32                                                  612.1Ki ± 36%             612.9Ki ± 29%       ~ (p=0.205 n=8)
FolderListAndSearch/list_all_inherited_dashboards_with_nested_folders_feature_enabled-32                                   21.37Mi ±  9%             23.37Mi ± 17%       ~ (p=0.382 n=8)
FolderListAndSearch/search_for_pattern_with_nested_folders_feature_enabled-32                                              307.4Ki ± 11%             291.5Ki ±  7%       ~ (p=0.105 n=8)
FolderListAndSearch/search_for_specific_dashboard_nested_folders_feature_enabled-32                                        78.37Ki ±  2%             74.09Ki ±  2%  -5.46% (p=0.000 n=8)
FolderListAndSearch/get_root_folders_with_nested_folders_feature_disabled-32                                               1.027Mi ±  0%             1.028Mi ±  0%       ~ (p=0.442 n=8)
FolderListAndSearch/list_all_dashboards_with_nested_folders_feature_disabled-32                                            21.50Mi ±  5%             23.17Mi ±  0%  +7.78% (p=0.007 n=8)
FolderListAndSearch/search_specific_dashboard_with_nested_folders_feature_disabled-32                                      77.55Ki ±  2%             73.26Ki ±  1%  -5.53% (p=0.000 n=8)
geomean                                                                                                                    1.156Mi                   1.154Mi        -0.10%

                                                                                         │ postgres_logs_grafana_grafana_125182_1_13.log │ postgres_logs_grafana_grafana_125193_1_13.log │
                                                                                         │                   allocs/op                   │        allocs/op         vs base              │
FolderListAndSearch/get_root_folders_with_nested_folders_feature_enabled-32                                                  152.7k ± 0%               152.7k ± 0%       ~ (p=0.328 n=8)
FolderListAndSearch/get_subfolders_with_nested_folders_feature_enabled-32                                                    15.54k ± 0%               15.54k ± 0%       ~ (p=0.485 n=8)
…